Decouple (make independent) pybind module counts_and_probs from core CUNQA functionalities by making probabilities a free function
